+++ src/Service.Liquidity.Portfolio.Grpc/Models/GetProjectionRequest.cs
@@ -8,6 +8,6 @@ namespace Service.Liquidity.Portfolio.Grpc.Models
         [DataMember(Order = 1)] public string BrokerId { get; set; }
         [DataMember(Order = 2)] public string FromAsset { get; set; }
         [DataMember(Order = 3)] public string ToAsset { get; set; }
-        [DataMember(Order = 4)] public double FromVolume { get; set; }
+        [DataMember(Order = 4)] public decimal FromVolume { get; set; }
     }
 }
