+++ src/Service.Liquidity.Portfolio.Grpc/Models/UpdateBalanceRequest.cs
@@ -8,7 +8,7 @@ namespace Service.Liquidity.Portfolio.Grpc.Models
         [DataMember(Order = 1)] public string BrokerId { get; set; }
         [DataMember(Order = 2)] public string WalletName { get; set; }
         [DataMember(Order = 3)] public string Asset { get; set; }
-        [DataMember(Order = 4)] public double BalanceDifference { get; set; }
+        [DataMember(Order = 4)] public decimal BalanceDifference { get; set; }
         [DataMember(Order = 5)] public string Comment { get; set; }
         [DataMember(Order = 6)] public string User { get; set; }
     }
